+++ src/features/manage-night-study-applications/ui/AttendanceTable.tsx
@@ -3,9 +3,7 @@ import { useDebounce } from "@/shared/hooks/useDebounce";
 import QueryBoundary from "@/shared/ui/query-boundary";
 import {
   Dropdown,
-  FilledButton,
   Table,
-  useToast,
 } from "@b1nd/dodam-design-system/components";
 import { colors } from "@b1nd/dodam-design-system/colors";
 import { MagnifyingGlass } from "@b1nd/dodam-design-system/icons";
@@ -15,9 +13,9 @@ import {
   type AttendanceFilterStatus,
   useAttendanceTable,
 } from "../hooks/useAttendanceTable";
-import { downloadAttendanceSummaryExcel } from "../utils/night-study-excel";
 import AttendanceActionCell from "./AttendanceActionCell";
 import AttendanceSkeletonRows from "./AttendanceSkeletonRows";
+import NightStudyAttendanceExcelButton from "./NightStudyAttendanceExcelButton";
 
 const ATTENDANCE_FILTER_ITEMS = [
   { name: "전체 상태", value: "ALL" },
@@ -29,42 +27,6 @@ const ATTENDANCE_PERIOD_ITEMS = [
   { name: "심자 1", value: "1" },
   { name: "심자 2", value: "2" },
 ];
-
-const AttendanceExcelButton = ({
-  keyword,
-  period,
-}: {
-  keyword: string;
-  period: number;
-}) => {
-  const toast = useToast();
-  const [isDownloading, setIsDownloading] = useState(false);
-
-  const handleDownload = async () => {
-    try {
-      setIsDownloading(true);
-      await downloadAttendanceSummaryExcel({ keyword, period });
-    } catch (error) {
-      toast.error(
-        error instanceof Error ? error.message : "엑셀 파일 생성에 실패했습니다.",
-      );
-    } finally {
-      setIsDownloading(false);
-    }
-  };
-
-  return (
-    <FilledButton
-      role="assistive"
-      size="medium"
-      display="inline"
-      disabled={isDownloading}
-      onClick={handleDownload}
-    >
-      {isDownloading ? "다운로드 중.." : "엑셀 다운로드"}
-    </FilledButton>
-  );
-};
 
 const AttendanceTableData = ({
   keyword,
@@ -153,7 +115,7 @@ const AttendanceTable = () => {
           />
         </div>
         <div className="flex items-center gap-3 shrink-0">
-          <AttendanceExcelButton keyword={query} period={period} />
+          <NightStudyAttendanceExcelButton keyword={query} period={period} />
           <Dropdown
             items={ATTENDANCE_FILTER_ITEMS}
             value={attendStatus}
